fix size_t to int conversion for msvc

+++ src/kml/convenience/csv_parser.cc
@@ -107,7 +107,7 @@ CsvParserStatus CsvParser::SetSchema(const kmlbase::StringVector& csv_schema) {
     } else if (kmlbase::StringCaseEqual(this_col, "style-id")) {
       style_id_ = i;
     } else {
-      csv_schema_[i] = this_col;
+      csv_schema_[i] = static_cast<int>(this_col);
     }
   }
   if (lat_col_ == npos || lon_col_ == npos) {
@@ -151,7 +151,7 @@ CsvParserStatus CsvParser::CsvLineToPlacemark(
   }
   // Walk the actual line cols to handle non-strict mode.
   for (size_t i = 0; i < csv_line.size(); ++i) {
-    CsvSchema::const_iterator iter = csv_schema_.find(i);
+    CsvSchema::const_iterator iter = csv_schema_.find(static_cast<int>(i));
     if (iter != csv_schema_.end()) {
       AddExtendedDataValue(iter->second, csv_line[iter->first], placemark);
     }
